Remove 'dgpu' subcommand
The discrete GPU power cannot be set directly any more, instead the
kernel takes care of this itself via the runtime power-management
framework.

There shouldn't be any kernels with that interface in use any more, so
let's remove it. We may eventually re-introduce an adapted interface as
a convenience wrapper around the standard PCI sysfs attributes later on.